requirements.txt
# Stoic Wisdom Shorts Generator - Dependencies
# Install with: pip install -r requirements.txt

# Core
requests>=2.28.0
python-dotenv>=1.0.0

# Text-to-Speech
edge-tts>=6.1.0

# Audio Processing
pydub>=0.25.0
audioop-lts>=0.2.0; python_version>"3.12"

# Video Processing
moviepy==1.0.3
opencv-python-headless>=4.8.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.7

# PIL Text Rendering
Pillow>=10.0.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0

# Database
sqlalchemy>=2.0.0

# Logging
loguru>=0.7.0

# Scheduling (optional)
apscheduler>=3.10.0

# Timezone handling
pytz>=2023.3

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
